fix issue math.trunc() in FE and change int into bigdemical for quantity. Fix isue Pending order không được nhận snapshot giá/khuyến mãi/voucher/shipping từ client như source of truth. Combo availability không được dùng ProductVariant.stockQty; phải dùng batch sellable truth của component. Cancel/void invoice không được restore stock bằng heuristic FEFO/cost nếu thiếu allocation trace. Inventory valuation không được trộn variant.stockQty với avg cost tính từ batch predicate khác.

+++ NhaDanShop/src/main/java/com/example/nhadanshop/dto/ExcelPreviewResponse.java
@@ -37,7 +37,7 @@ public record ExcelPreviewResponse(
             String productCode,
             String variantCode,
             String productName,
-            Integer quantity,
+            BigDecimal quantity,
             BigDecimal unitCost,
             BigDecimal sellPrice,
             BigDecimal discountPercent,
